compilation fixes after modules update

+++ lib/issues.go
@@ -32,7 +32,7 @@ func CompareIssues(config cfg.Config, ghClient clients.GitHubClient, jiraClient 
 		return nil
 	}
 
-	ids := make([]int, len(ghIssues))
+	ids := make([]int64, len(ghIssues))
 	for i, v := range ghIssues {
 		ids[i] = v.GetID()
 	}
@@ -48,7 +48,7 @@ func CompareIssues(config cfg.Config, ghClient clients.GitHubClient, jiraClient 
 		found := false
 		for _, jIssue := range jiraIssues {
 			id, _ := jIssue.Fields.Unknowns.Int(config.GetFieldKey(cfg.GitHubID))
-			if int64(*ghIssue.ID) == id {
+			if *ghIssue.ID == id {
 				found = true
 				if err := UpdateIssue(config, ghIssue, jIssue, ghClient, jiraClient); err != nil {
 					log.Errorf("Error updating issue %s. Error: %v", jIssue.Key, err)
